requirements.txt
pandas==2.2.0
Pillow==10.2.0
pytest==7.4.3
scikit_learn==1.2.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.62.1
transformers==4.33.2
